# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI-powered E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:

    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating, newest]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [products]
      summary: Get product reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: addProductReview
      tags: [products]
      summary: Add a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /pricing/forecast:
    get:
      operationId: getPricingForecast
      tags: [pricing]
      summary: Get demand forecast and pricing suggestions
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pricing forecast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingForecastResponse"
  /pricing/adjust:
    post:
      operationId: adjustPrice
      tags: [pricing]
      summary: AI-powered price adjustment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustPriceBody"
      responses:
        "200":
          description: Price adjustment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAdjustmentResponse"
...
  /openai/conversations:
    get:
      operationId: listConversations
      tags: [openai]
      summary: List conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [openai]
      summary: Create a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /openai/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [openai]
      summary: Get messages in conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send message to conversation (SSE streaming)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: Streaming response (SSE)
          content:
            text/event-stream:
              schema:
                type: string
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        brand:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        specs:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
        demandScore:
          type: number
        createdAt:
          type: string
      required: [id, name, price, category, stock, rating, reviewCount]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        brand:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        specs:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
      required: [name, price, category, stock]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        brand:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        specs:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        author:
          type: string
        rating:
          type: integer
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
      required: [id, productId, author, rating, body]
    CreateReviewBody:
      type: object
      properties:
        author:
          type: string
        rating:
          type: integer
        title:
          type: string
        body:
          type: string
      required: [author, rating, body]
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        negotiatedPrice:
          type: number
      required: [id, productId, quantity]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        userId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        shippingAddress:
          type: string
        createdAt:
          type: string
      required: [id, sessionId, status, items, total]
    CreateOrderBody:
      type: object
      properties:
        sessionId:
          type: string
        userId:
          type: integer
        shippingAddress:
          type: string
        paymentMethod:
          type: string
          enum: [card, upi, cod]
      required: [sessionId]
...
    PricingForecastResponse:
      type: object
      properties:
        productId:
          type: integer
        currentPrice:
          type: number
        suggestedPrice:
          type: number
        demandLevel:
          type: string
          enum: [low, medium, high, surge]
        demandScore:
          type: number
        forecast:
          type: string
        alerts:
          type: array
          items:
            type: string
        stockStatus:
          type: string
      required: [productId, currentPrice, suggestedPrice, demandLevel, demandScore, forecast, stockStatus]
    AdjustPriceBody:
      type: object
      properties:
        productId:
          type: integer
        reason:
          type: string
      required: [productId]
    PriceAdjustmentResponse:
      type: object
      properties:
        productId:
          type: integer
        oldPrice:
          type: number
        newPrice:
          type: number
        reason:
          type: string
        changePercent:
          type: number
      required: [productId, oldPrice, newPrice, reason, changePercent]
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
      required: [id, title]
    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
      required: [title]
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required: [id, conversationId, role, content]
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
...