# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";

export interface CreateOrderBody {
  sessionId: string;
  userId?: number;
  shippingAddress?: string;
  paymentMethod?: CreateOrderBodyPaymentMethod;
}