# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  sessionId: string;
  userId?: number;
  status: OrderStatus;
  items: CartItem[];
  total: number;
  shippingAddress?: string;
  createdAt?: string;
}