# lib/api-zod/src/generated/types/updateProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProductBodySpecs } from "./updateProductBodySpecs";

export interface UpdateProductBody {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  brand?: string;
  imageUrl?: string;
  stock?: number;
  specs?: UpdateProductBodySpecs;
  tags?: string[];
}